fix: TickTick sync schema compatibility and robust data handling
Schema Fixes:
- Change Project.sort_order from INTEGER to BIGINT (values exceed 2B)
- Change Task.sort_order from INTEGER to BIGINT (values like -2199023255552)
- TickTick uses large sort_order values that don't fit in PostgreSQL INTEGER

Data Handling Improvements:
- Add _extract_first_reminder() method to handle both V1 and V2 API formats
- Add isinstance() checks for tags, pomodoroSummaries, focusSummaries
- Prevent 'str' object has no attribute 'get' errors

Files Changed:
- app/models/project.py: BigInteger import and sort_order type
- app/models/task.py: BigInteger import and sort_order type
- app/services/ticktick.py: Robust reminder extraction and type guards

+++ backend/app/models/project.py
@@ -1,7 +1,7 @@
 """
 Project model for TickTick project integration.
 """
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.core.database import Base
@@ -24,7 +24,7 @@ class Project(Base):
     # Project Info
     name = Column(String(500), nullable=False)
     color = Column(String(50), nullable=True)  # Hex color code
-    sort_order = Column(Integer, default=0, nullable=False)
+    sort_order = Column(BigInteger, default=0, nullable=False)
     is_archived = Column(Boolean, default=False, nullable=False)
 
     # Timestamps
